Migrate project to pnpm
I was unable to use the project locally through pnpm link from other
project because of react library installed at two different places which
is why I tried to use the `file:` scheme as recommended on pnpm website.
However, it only works if the project uses the same package manager.